Projects: extract order as help type, adapt references
