+++ src/app/projects/page.tsx
@@ -5,7 +5,7 @@ import { useSearchParams } from 'next/navigation'
 import { useQuery } from '@tanstack/react-query'
 import { Center, Pagination } from '@mantine/core'
 
-import { Order, ProjectQueryOptions } from '@/services/projects'
+import { ProjectQueryOptions } from '@/services/projects'
 import { CurrentUserQueryOptions } from '@/services/currentUser'
 import { DepartmentQueryKey, ResearchDepartments } from '@/services/departments'
 import { Facilities } from '@/services/facilities'
@@ -16,6 +16,7 @@ import { Users } from '@/services/users'
 import Filter from '@/components/Common/Filter/Filter'
 import ProjectFilterContent from '@/components/Project/ProjectFilterContent'
 import ProjectsList from '@/components/Project/ProjectsList'
+import { Order } from '@/entities/HelpTypes/Order'
 import { ProjectSortAttribute, RequestState } from '@/entities/ProjectInList'
 
 const ProjectsPage: NextPage = () => {
